+++ README.md
@@ -149,6 +149,9 @@ mutually exclusive.
                  causes the iODBC driver to be built.
 ```
 
+By default, the ODBC driver will be installed as /usr/local/lib/odbc/libmdbodbc.so,
+with a Unicode-capable driver at /usr/local/lib/odbc/libmdbodbcW.so.
+
 A list of general options is available in the [INSTALL](./INSTALL) file, and
 `configure --help` will give you the list of mdbtools specific options.
